Ported the ability to generate a KeyPair from an address and an account number. I also added a unit test to SeedTest to test generating a keyPair from an address with checks generated by the JS library.
One area of extension would be to eliminate the notion of sending in a seedStr when generating a keyPair from an address or account number and having this just be a setable field within the Seed class. Given the nature of how the account class interacts with this class thought I felt that decision should be left up to someone else.